#2576 Bug fix: look description
When users are chosen and then we go to another section (Team for
example) and then go back to Reports section, users’ status (chosen) is
lost but the reports gotten are still there. If we clicked calculate,
the report is not cleaned. We had to double click on users and click
Calculate in order to clean the report. Now users’ statuses are saved
to Store and going to another sections doesn’t affect it.